+++ src/TorchlightServiceProvider.php
@@ -5,7 +5,6 @@
 
 namespace Torchlight;
 
-use Torchlight\Blade\BladeManager;
 use Torchlight\Blade\CodeComponent;
 use Torchlight\Commands\Install;
 use Illuminate\Support\ServiceProvider;
@@ -14,6 +13,8 @@ class TorchlightServiceProvider extends ServiceProvider
 {
     public function boot()
     {
+        $this->app->singleton(Manager::class);
+
         if ($this->app->runningInConsole()) {
             $this->commands([
                 Install::class,
@@ -24,12 +25,11 @@ class TorchlightServiceProvider extends ServiceProvider
             __DIR__ . '/../config/torchlight.php' => config_path('torchlight.php')
         ], 'config');
 
-        if (config('torchlight.blade_components')) {
+        if (Torchlight::config('torchlight.blade_components')) {
             $this->loadViewComponentsAs('torchlight', [
                 'code' => CodeComponent::class
             ]);
         }
-
     }
 
     public function register()