+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.0 - 2021-05-21
+
+- Add `Torchlight` facade
+- Add ability to set the cache implementation.
+- Add ability to abstract config from Laravel's `config` helper.
+
+
 ## 0.2.1 - 2021-05-20
 
 - Add `Block::generateIdsUsing`